Input: wacom - add Intuos5 Touch Ring LED support
commit 9b5b95dd516a13d53ecf9217672d2116f05097bc upstream.

The Touch Ring LEDs on Intuos5 tablets use a different report
format which supports only 4 levels of brightness. We remap
the 7-bit value obtained from sysfs to an appropriate value
for the tablet. Control of the crop mark LEDs (new to the I5)
is left for a later patch.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Dmitry Torokhov <dtor@mail.ru>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Gbp-Pq: Topic features/all/wacom
Gbp-Pq: Name 0022-Input-wacom-add-Intuos5-Touch-Ring-LED-support.patch