be2net: Fix EEH error reset before a flash dump completes
commit eeb7fc7bc095546b21188e8e076a59bce73f9ca6 upstream.

An EEH error can cause the FW to trigger a flash debug dump.
Resetting the card while flash dump is in progress can cause it not to recover.
Wait for it to finish before letting EEH flow to reset the card.

Signed-off-by: Sathya Perla <Sathya.Perla@emulex.com>
Signed-off-by: Somnath Kotur <somnath.kotur@emulex.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Gbp-Pq: Topic features/all/be2net
Gbp-Pq: Name 0057-be2net-Fix-EEH-error-reset-before-a-flash-dump-compl.patch