virtio-scsi: split locking per vq
commit 139fe45abc2234b20fd5ecbcb7ea6d3688fed5e5 upstream.

Keep a separate lock for each virtqueue.  While not particularly
important now, it prepares the code for when we will add support
for multiple request queues.  It is also more tidy as soon as
we introduce a separate lock for the sglist.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
[bwh: Backported to 3.2: s/virtqueue_add_buf/&_gfp/]

Gbp-Pq: Topic features/all/virtio_scsi
Gbp-Pq: Name 0004-SCSI-virtio-scsi-split-locking-per-vq.patch