wacom: do not request tablet data on MT Tablet PC pen interface
When in commit 1963518b9b1b (Input: wacom - add 0xE5 (MT device)
support, 2012-04-29) the driver stopped asking for multitouch tablet
data on the pen interface of a tablet PC, as a side effect we started
executing the "else" to that if statement.  Oops.

This is needed for the 0xE6 tablet in the Thinkpad x220t to be usable
again.  Meanwhile the 0xE3 works fine without this.  Not sure why. -jn

[jn: extracted from a larger commit in the input-wacom repository,
 with new description]
[bwh: Part of commit ea2e60244573a9204c8cee9b4fb181106784c617 upstream]

Signed-off-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Gbp-Pq: Topic features/all/wacom
Gbp-Pq: Name 0029-wacom-do-not-request-tablet-data-on-MT-Tablet-PC-pen.patch