apparmor: remove advertising the support of network rules from compat iface
commit 873143ceca69a2e54e7face1be49ad6b5514525d in
git://git.kernel.org/pub/scm/linux/kernel/git/jj/linux-apparmor

The interface compatibility patch was advertising support of network rules,
however this is not true if the networking patch is not applied. Move
advertising of network rules into a third patch that can be applied if
both the compatibility and network patches are applied.

Signed-off-by: John Johansen <john.johansen@canonical.com>

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name apparmor-remove-advertising-the-support-of-network-r.patch