codel: refine one condition to avoid a nul rec_inv_sqrt
commit 2359a47671fc4fb0fe5e9945f76c2cb10792c0f8 upstream.

One condition before codel_Newton_step() was not good if
we never left the dropping state for a flow. As a result
rec_inv_sqrt was 0, instead of the ~0 initial value.

codel control law was then set to a very aggressive mode, dropping
many packets before reaching 'target' and recovering from this problem.

To keep codel_vars_init() as efficient as possible, refine
the condition to make sure rec_inv_sqrt initial value is correct

Many thanks to Anton Mich for discovering the issue and suggesting
a fix.

Reported-by: Anton Mich <lp2s1h@gmail.com>
Signed-off-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Gbp-Pq: Topic features/all/codel
Gbp-Pq: Name 0008-codel-refine-one-condition-to-avoid-a-nul-rec_inv_sq.patch