igb: Do not use header split, instead receive all frames into a single buffer
This change makes it so that we no longer use header split.  The idea is to
reduce partial cache line writes by hardware when handling frames larger
then header size.  We can compensate for the extra overhead of having to
memcpy the header buffer by avoiding the cache misses seen by leaving an
full skb allocated and sitting on the ring.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
Tested-by: Aaron Brown <aaron.f.brown@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
[bwh: Backported to wheezy:
 - Keep using alloc_page(), not __skb_alloc_page()
 - Don't test for NETIF_F_RXALL
 - skb_add_rx_frag() doesn't have a truesize parameter]

Gbp-Pq: Topic features/all/igb
Gbp-Pq: Name 0058-igb-Do-not-use-header-split-instead-receive-all-fram.patch