igb: Map entire page and sync half instead of mapping and unmapping half pages
This change makes it so that we map the entire page and just sync half of
it for the device at a time.  The advantage to this approach is that we can
avoid the locking on map/unmap seen in many IOMMU implementations.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
Tested-by: Aaron Brown <aaron.f.brown@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
[bwh: Backported to wheezy:
 - Keep using alloc_page(), not __skb_alloc_page()
 - skb_add_rx_frag() doesn't have a truesize parameter]

Gbp-Pq: Topic features/all/igb
Gbp-Pq: Name 0060-igb-Map-entire-page-and-sync-half-instead-of-mapping.patch