igb: Fix SerDes autoneg flow control.
This patch enables flow control to be set in SerDes autoneg mode.  This is
done the way it is done for copper, but relies on a different set of register/bit
checks since this is all done within the MAC registers.

Signed-off-by: Carolyn Wyborny <carolyn.wyborny@intel.com>
Tested-by: Aaron Brown <aaron.f.brown@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>

Gbp-Pq: Topic features/all/igb
Gbp-Pq: Name 0088-igb-Fix-SerDes-autoneg-flow-control.patch