rtnetlink: avoid ABI change in 3.2.34
Commit d318a127e273716c9531fe70d497ca24db4c0bf1 ('rtnetlink: Fix
problem with buffer allocation', commit
115c9b81928360d769a76c632bae62d15206a94a upstream) added a parameter
to the optional 'calcit' callback for rtnetlink operations.  There is
only one in-tree implementation and it's in the same file as the
caller, so call it directly with the extra argument when necessary.

Also, the ifla_policy array grew, but this is backward-compatible so
hide the change from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name rtnetlink-avoid-ABI-change-in-3.2.34.patch