staging: line6: drop variax model_data field
commit 81804c54db877b78b12bd1cf4f4e0f7bf6c86d72 upstream.

The sysfs attrs to access the model dump have been removed so it's safe
to drop the model_data field.  The next step will be to simplify the
startup process since we no longer need to fetch this data via dump
requests, but that will be done in a later patch.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Gbp-Pq: Topic features/all/line6
Gbp-Pq: Name 0103-staging-line6-drop-variax-model_data-field.patch